Split code out of the massive mod.rs
This should make it easier to work on an encourage more smaller more
deliberate interface creation

This fixes #61
