+++ pb-commit-message-lints/src/lints/cmd/mod.rs
@@ -1,0 +1,5 @@
+mod lint;
+mod set_status;
+
+pub use lint::lint;
+pub use set_status::set_status;
