+++ pb-commit-message-lints/src/lints/lib/mod.rs
@@ -1,5 +1,13 @@
 mod commit_message;
+mod lint;
 mod lints;
 
 pub use commit_message::CommitMessage;
+pub use lint::Lint;
 pub use lints::Lints;
+pub use problem::{Code, Problem};
+
+mod duplicate_trailers;
+mod missing_jira_issue_key;
+mod missing_pivotal_tracker_id;
+mod problem;
